+++ CHANGELOG.rst
@@ -8,6 +8,9 @@ v34.6.3 (unreleased)
   for fetching Docker images using skopeo through ``run_command_safely`` calls.
   https://github.com/nexB/scancode.io/issues/1257
 
+- Fix an issue in the d2d JavaScript mapper.
+  https://github.com/nexB/scancode.io/pull/1274
+
 v34.6.2 (2024-06-18)
 --------------------